# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: TheChickenMan B2B Chicken Marketplace API
  version: 1.0.0
  description: Authoritative RESTful API for TheChickenMan commercial poultry platform.
servers:
  - url: /api/v1
    description: Production & Local API Gateway
paths:
...
  /inventory/batches:
    get:
      summary: List supplier production batches
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Batch list }
    post:
      summary: Create production batch
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Batch created }
  /inventory:
    get:
      summary: View inventory ledger balances
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Inventory records }
    post:
      summary: Add stock (STOCK_IN)
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Stock added }
...
  /orders:
    get:
      summary: List orders (filtered by authenticated role)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of orders }
    post:
      summary: Checkout cart and create order (reserves inventory)
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Order created and inventory reserved }
  /orders/{id}:
    get:
      summary: Get order details and status history
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Order details }
  /orders/{id}/accept:
    post:
      summary: Supplier accepts order (full or partial)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Order confirmed }
  /orders/{id}/reject:
    post:
      summary: Supplier rejects order
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Order rejected and stock released }
  /orders/{id}/status:
    patch:
      summary: Update order status (Processing, Packed, Dispatched)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Order status updated }
  /payments/intent:
    post:
      summary: Create payment intent
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Payment intent created }
  /payments/verify:
    post:
      summary: Verify payment completion
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Payment verified and order marked paid }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT